// src/modules/kaspiapi/dto/check.request.kaspi.dto.ts
export interface CheckRequestKaspiDto {
  /**
   * check – запрос на проверку состояния абонента/номера заказа
   *
   *
   * pay – запрос на пополнение баланса абонента/номера заказа
   * */
  command: 'check';

  /**
   * внутренний номер запроса в системе Kaspi.kz
   * */
  txn_id: string;

  /**
   * идентификатор абонента/номера заказа в информационной системе провайдера
   * */
  account: string;

  /**
   * сумма к зачислению на лицевой счет абонента
   * (в запросе check значение в переменной sum является фиктивным,
   * передается с терминала по умолчанию и его обрабатывать не нужно)
   */
  sum: number;

  /**
   * Дополнительные поля, передаваемые провайдером - строки, содержащие буквы, цифры и спецсимволы.
   *
   *
   * (NOTE) Парсить с помощью pipe:
   *
   *
   * (EXAMPLE) data1,data2,…,dataN
   *
   * */
  device_id: string;

  /**
   * payment comment
   * */
  comment: string;
}